requirements.txt
aws-cdk-lib==2.99.1
constructs==10.2.70 #>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.99.1a0
boto3==1.28.59
